requirements.txt
# Core dependencies
crewai>=0.70.0
qdrant-client>=1.11.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
# Embeddings
sentence-transformers>=3.0.0
# API and LLM
groq>=0.4.0
anthropic>=0.34.0
# Frontend
streamlit>=1.38.0
plotly>=5.18.0
# Utilities
pydantic>=2.0.0
python-dateutil>=2.8.2
tqdm>=4.66.0
scikit-learn>=1.3.0